Multiple Receivers Support
Now you can send your IP to multiple mails. Add the mail you want in a python array format in the config.json. An example is given in the fresh config.json, so if you want to see that delete your config file and recompile it)
I've also misspelled "receiver" every time in the code, now it's spelled right.